+++ Model/Config/Source/Allowwedpayment.php
@@ -40,6 +40,7 @@ class Allowwedpayment implements \Magento\Framework\Option\ArrayInterface
             ['value' => 'paypal', 'label' => __('PayPal')],
             ['value' => 'resurs', 'label' => __('Resurs Bank')],
             ['value' => 'vipps', 'label' => __('Vipps')],
+            ['value' => 'googlepay', 'label' => __('Google Pay')],
         ];
     }
 }