+++ Model/ConfigProvider.php
@@ -48,6 +48,7 @@ class ConfigProvider implements ConfigProviderInterface
             'resurs_payment_icons' => $this->_layout->createBlock('Radarsofthouse\Reepay\Block\Paymenticons')->setTemplate('Radarsofthouse_Reepay::resurs_payment_icons.phtml')->toHtml(),
             'vipps_payment_icons' => $this->_layout->createBlock('Radarsofthouse\Reepay\Block\Paymenticons')->setTemplate('Radarsofthouse_Reepay::vipps_payment_icons.phtml')->toHtml(),
             'forbrugsforeningen_payment_icons' => $this->_layout->createBlock('Radarsofthouse\Reepay\Block\Paymenticons')->setTemplate('Radarsofthouse_Reepay::forbrugsforeningen_payment_icons.phtml')->toHtml(),
+            'googlepay_payment_icons' => $this->_layout->createBlock('Radarsofthouse\Reepay\Block\Paymenticons')->setTemplate('Radarsofthouse_Reepay::googlepay_payment_icons.phtml')->toHtml(),
             'saved_credit_cards' => $this->_layout->createBlock('Radarsofthouse\Reepay\Block\SavedCreditCards')->setTemplate('Radarsofthouse_Reepay::saved_credit_cards.phtml')->toHtml(),
             "reepay_payment_instructions" => $this->_scopeConfig->getValue('payment/reepay_payment/instructions',\Magento\Store\Model\ScopeInterface::SCOPE_STORE,$store_id),
             "reepay_applepay_instructions" => $this->_scopeConfig->getValue('payment/reepay_applepay/instructions',\Magento\Store\Model\ScopeInterface::SCOPE_STORE,$store_id),
@@ -60,7 +61,8 @@ class ConfigProvider implements ConfigProviderInterface
             "reepay_vipps_instructions" => $this->_scopeConfig->getValue('payment/reepay_vipps/instructions',\Magento\Store\Model\ScopeInterface::SCOPE_STORE,$store_id),
             "reepay_forbrugsforeningen_instructions" => $this->_scopeConfig->getValue('payment/reepay_forbrugsforeningen/instructions',\Magento\Store\Model\ScopeInterface::SCOPE_STORE,$store_id),
             "reepay_viabill_instructions" => $this->_scopeConfig->getValue('payment/reepay_viabill/instructions',\Magento\Store\Model\ScopeInterface::SCOPE_STORE,$store_id),
-            "reepay_paypal_instructions" => $this->_scopeConfig->getValue('payment/reepay_paypal/instructions',\Magento\Store\Model\ScopeInterface::SCOPE_STORE,$store_id)
+            "reepay_paypal_instructions" => $this->_scopeConfig->getValue('payment/reepay_paypal/instructions',\Magento\Store\Model\ScopeInterface::SCOPE_STORE,$store_id),
+            "reepay_googlepay_instructions" => $this->_scopeConfig->getValue('payment/reepay_googlepay/instructions',\Magento\Store\Model\ScopeInterface::SCOPE_STORE,$store_id)
         ];
     }
 }